Move integration test to tests dir
